Scouter AI Done! + Parasite Faction Beginnings
Scouter will wander with bias towards player. When a new player-made object is found far enough from existing player base or if there is no previous base, the scouter will avoid all danger and immediately try to research this object. Once the threshold of 100 is hit the scouter's job will be done and it will flee from the player and despawn far away. If the parasite is ever hit it will immediately fight back as a self defense measure. It will cycle between 2 moves, a serious of 3 biting attacks, and one launching attack. The scouter will resume wandering and researching once the player leaves the immediate vicinity.
Next up: Parasite Skirmishers and Soldiers