Ether Shard mechanic added
Another safety net for the player. Upon dying without adrenaline the player and enemy will be fully healed and sent to the ether. There they must compete in a fair and square 1v1. The only way to get back to reality is to win the duel. Dying in the ether or without a full shard charge results in game over, and save file deletion. Shard can be charged by slaying enemies.