FIXED CAMERA AND PLAYER MOVEMENT!!
Previously we were updating everything visually in fixedupdate, which caused the game to basically only update at 50hz!! The camera now updates in lateupdate, and now the player moves with velocity instead of changing position. The player also has interpolation enabled on the rb so that it can update visually every regular frame. Great update!!!